Creat private chat room
