+++ src/PrivateChatRoom.css
@@ -1,0 +1,88 @@
+.App {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  text-align: center;
+  max-width: 414px;
+  height: 100%;
+  margin: auto;
+}
+
+@media (min-width: 700px) {
+  .App {
+    max-width:90%;
+  }
+}
+@media (min-width: 950px) {
+  .App {
+    max-width:100%;
+  }
+}
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+.App-header {
+  flex: 1;
+  background-color: #3d14f5;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  color: rgb(0, 0, 0);
+}
+
+.App-chatroom 
+{
+  flex: 12;
+  background-color:#2D2C67
+}
+
+.App-chatroom-text 
+{
+  margin: 8px;
+  
+  background-color: rgb(144, 218, 101);
+  color : rgb(48, 100, 100);
+  font-size: 20px;
+}
+
+.App-textbox {
+  flex: 1;
+  justify-content: center;
+  align-items: stretch;
+  display: flex;
+}
+
+.App-textbox-input {
+  flex: 5;
+  height: 100%;
+  box-sizing: border-box;
+  image-rendering: auto;
+}
+
+.App-textbox-send {
+  flex: 1;
+  height: 100%;
+  background-color: rgb(252, 8, 0);
+  border: none;
+  font-size: 15px;
+  cursor: pointer;
+  padding: 8px 5px;
+  vertical-align: middle;
+  justify-content: center;
+  align-items: center;
+  display: flex;
+}
+.btn {
+  background-color: DodgerBlue;
+  grid: flex;
+  border: none; 
+  color: white;
+  padding: 8px 5px; /* Some padding */
+  font-size: 15px; /* Set a font size */ 
+}
+.btn:hover{
+  color:RoyalBlue;
+}
